Current GitHub Files/Decryption/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision==2.3.1